feat(cli): test introspection and migration Rust panics with no interactive mode
